move activity tracking to 'activity' namespace
